requirements.txt
fastapi
uvicorn[standard]
pydantic-settings
azure-storage-blob
motor
python-dotenv
aiofiles